refactor: act request and func
